CHERI-C++: pass structs and classes indirectly (for now).
Currently, when passing a C++ struct or class by value, Clang will pass the
fields as arguments (including capabilities). This is fine when there is no
inheritance. If a superclass contains a capability, this capability will be
turned into { i64, i64, i64, i64 } before each individual i64 is passed as an
argument and then the capability is reconstructed in the callee. This leads to
losing the tag.

For the time being we pass such C++ structs/classes indirectly (but set the
byval LLVM attribute on the parameter).

A TODO is to look at properly coercing objects when a superclass contains
capabilities but this will require some thinking, so leaving for later.